Runtime/RTAcpi*: Parser updates, the code is now able to parse vbox.dsl successfully and generate an AST, bugref:10733 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108187 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/acpi/acpi.cpp
@@ -1740,7 +1740,7 @@ static int rtAcpiResourceAddQWordAddressRange(PRTACPIRESINT pThis, uint8_t bType
                                               uint64_t u64Granularity, uint64_t u64Length,
                                               const char *pszRsrcSrc, uint8_t bRsrcIndex)
 {
-    size_t cchRsrcSrc = pszRsrcSrc ? strlen(pszRsrcSrc) + 2 : 0;
+    uint32_t cchRsrcSrc = pszRsrcSrc ? (uint32_t)strlen(pszRsrcSrc) + 2 : 0;
     uint8_t *pb = rtAcpiResBufEnsureSpace(pThis, 3 + 43 + cchRsrcSrc);
     if (!pb)
         return VERR_NO_MEMORY;
